scubotgo/requirements.txt
Django==4.2.7
line-bot-sdk==1.19.0
gspread